Handle changes in display mode and scale the render accordingly
This implements handling in the engine for changes in display resolution, fullscreen, and window resizing.

Scene and entity drawing was also changed - they no longer `blit` directly on the display surface. Instead, they now `blit` to the render layer surface, stored as an attribute in the game manager and passed with the draw calls.